rough cut push/pull to GCS
I was hoping to implement push/pull by calling the `gsutil` CLI but I'm not finding a way to make it create the psuedo directory placeholders needed for `gcsfuse` to be fast.

Also gsutil can copy files in parallel but only to a single target base directory, that is, not a list of (source:target) pairs. That's still useful for each recursive directory tree and for sibling directories if the code organizes them that way, but it's not max parallelism.

So we probably need to make it call the API.