clean up StackDriver logging
Clean up StackDriver logging: Flush and remove top level background-transport logging handlers so the last messages get to the server and won't raise `RuntimeError('cannot schedule new futures after shutdown')`.

StackDriver should be out of the loop after that but there's no documented API for this so it's not guaranteed.

Name the StackDriver `fireworker` instead of the default, `python`.